# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Career Copilot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /resumes:
    get:
      operationId: listResumes
      tags: [resumes]
      summary: List all resumes for the current user
      responses:
        "200":
          description: List of resumes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resume"
    post:
      operationId: createResume
      tags: [resumes]
      summary: Upload or create a resume
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeInput"
      responses:
        "201":
          description: Created resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
  /resumes/{id}:
    get:
      operationId: getResume
      tags: [resumes]
      summary: Get a single resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
        "404":
          description: Not found
    delete:
      operationId: deleteResume
      tags: [resumes]
      summary: Delete a resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /resumes/{id}/analyze:
    post:
      operationId: analyzeResume
      tags: [resumes]
      summary: Run AI analysis on a resume (ATS score, suggestions, skill gaps)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeAnalysis"
  /resumes/{id}/analysis:
    get:
      operationId: getResumeAnalysis
      tags: [resumes]
      summary: Get the latest analysis for a resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeAnalysis"
        "404":
          description: No analysis yet
  /resumes/{id}/improve:
    post:
      operationId: getResumeImprovements
      tags: [resumes]
      summary: Get detailed resume improvement suggestions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Improvement suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeImprovements"

  # ── Job Descriptions ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    Resume:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        content:
          type: string
        atsScore:
          type: ["number", "null"]
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - content
        - isDefault
        - createdAt
        - updatedAt
    ResumeInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        isDefault:
          type: boolean
      required:
        - title
        - content
    ResumeAnalysis:
      type: object
      properties:
        id:
          type: number
        resumeId:
          type: number
        atsScore:
          type: number
        overallFeedback:
          type: string
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        keywordsSuggested:
          type: array
          items:
            type: string
        skillsFound:
          type: array
          items:
            type: string
        skillsMissing:
          type: array
          items:
            type: string
        formattingScore:
          type: number
        contentScore:
          type: number
        impactScore:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - resumeId
        - atsScore
        - overallFeedback
        - strengths
        - weaknesses
        - keywordsSuggested
        - skillsFound
        - skillsMissing
        - formattingScore
        - contentScore
        - impactScore
        - createdAt
    ResumeImprovements:
      type: object
      properties:
        resumeId:
          type: number
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/ImprovementSuggestion"
        rewrittenSections:
          type: array
          items:
            $ref: "#/components/schemas/RewrittenSection"
      required:
        - resumeId
        - suggestions
        - rewrittenSections
    ImprovementSuggestion:
      type: object
      properties:
        section:
          type: string
        issue:
          type: string
        suggestion:
          type: string
        priority:
          type: string
      required:
        - section
        - issue
        - suggestion
        - priority
    RewrittenSection:
      type: object
      properties:
        section:
          type: string
        original:
          type: string
        improved:
          type: string
      required:
        - section
        - original
        - improved

    # ── Job Descriptions ───────────────────────────────────────────────────
...