# lib/api-zod/src/generated/types/improvementSuggestion.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AI Career Copilot API
 * OpenAPI spec version: 0.1.0
 */

export interface ImprovementSuggestion {
  section: string;
  issue: string;
  suggestion: string;
  priority: string;
}